don't create fhir resource and ehr if user is not using a SOF app
We need to check if the registration event handler is performing smart on fhir related tasks only if the registered user is registering for a smart on fhir app. this initial implementation checks for launch/patient scope. since this scope is allowed only for smart on fhir apps in our keycloak config, it ensures that it is a smart on fhir app user who is registering.